pyproject.toml
[project]
name = "tds-project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "async-timeout==4.0.3",
    "av==16.0.1",
    "bottleneck==1.6.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "coloredlogs==15.0.1",
    "contourpy==1.3.2",
    "cryptography==46.0.3",
    "ctranslate2==4.6.1",
    "cycler==0.12.1",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "duckdb>=1.4.2",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.1",
    "fastapi==0.121.3",
    "faster-whisper==1.2.1",
    "filelock==3.20.0",
    "flatbuffers==25.9.23",
    "fonttools==4.60.1",
    "fsspec==2025.10.0",
    "fuzzywuzzy[speedup]>=0.18.0",
    "geopy>=2.4.1",
    "google-api-core==2.28.1",
    "google-auth==2.43.0",
    "google-cloud-aiplatform==1.128.0",
    "google-cloud-bigquery==3.38.0",
    "google-cloud-core==2.5.0",
    "google-cloud-resource-manager==1.15.0",
    "google-cloud-storage==3.6.0",
    "google-crc32c==1.7.1",
    "google-genai==1.52.0",
    "google-resumable-media==2.8.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.4",
    "grpc-google-iam-v1==0.14.3",
    "grpcio==1.76.0",
    "grpcio-status==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.1.5",
    "humanfriendly==10.0",
    "idna==3.11",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.9",
    "langchain==1.0.8",
    "langchain-classic==1.0.0",
    "langchain-core==1.1.0",
    "langchain-google-vertexai==3.0.3",
    "langchain-openai==1.0.3",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.46",
    "matplotlib==3.10.7",
    "mpmath==1.3.0",
    "networkx>=3.4.2",
    "numexpr==2.14.1",
    "numpy==2.2.6",
    "onnxruntime==1.23.2",
    "openai==2.8.1",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pdfminer-six==20251107",
    "pdfplumber==0.11.8",
    "pillow==12.0.0",
    "playwright==1.56.0",
    "proto-plus==1.26.1",
    "protobuf==6.33.1",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pyee==13.0.0",
    "pyparsing==3.2.5",
    "pypdfium2==5.1.0",
    "pytesseract==0.3.13",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "shapely==2.1.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tqdm==4.67.1",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "validators==0.35.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
